+++ 10/lib/presentation/screens/ui_controls/ui_controls_screen.dart
@@ -1,0 +1,119 @@
+import 'package:flutter/material.dart';
+
+class UiControlsScreen extends StatelessWidget {
+  static const name = 'ui_controls_screen';
+
+  const UiControlsScreen({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: AppBar(title: Text("UI Controls")),
+      body: _UiControlsView(),
+    );
+  }
+}
+
+class _UiControlsView extends StatefulWidget {
+  const _UiControlsView();
+
+  @override
+  State<_UiControlsView> createState() => _UiControlsViewState();
+}
+
+enum Transportation { car, plane, boat, submarine }
+
+class _UiControlsViewState extends State<_UiControlsView> {
+  bool isDeveloper = true;
+  Transportation selectedTransportation = Transportation.car;
+  bool wantsBreakfast = false;
+  bool wantsLunch = false;
+  bool wantsDinner = false;
+
+  @override
+  Widget build(BuildContext context) {
+    return ListView(
+      physics: ClampingScrollPhysics(),
+      children: [
+        SwitchListTile(
+          title: Text("Developer Mode"),
+          subtitle: Text("Controles Adicionales"),
+          value: isDeveloper,
+          onChanged: (value) => setState(() {
+            isDeveloper = !isDeveloper;
+          }),
+        ),
+
+        ExpansionTile(
+          title: Text("Vehiculo de transporte"),
+          subtitle: Text("$selectedTransportation"),
+          children: [
+            RadioListTile(
+              title: Text("By Car"),
+              subtitle: Text("Viajar por carro"),
+              value: Transportation.car,
+              groupValue: selectedTransportation,
+              onChanged: (value) => setState(() {
+                selectedTransportation = Transportation.car;
+              }),
+            ),
+
+            RadioListTile(
+              title: Text("By boat"),
+              subtitle: Text("Viajar por boat"),
+              value: Transportation.boat,
+              groupValue: selectedTransportation,
+              onChanged: (value) => setState(() {
+                selectedTransportation = Transportation.boat;
+              }),
+            ),
+
+            RadioListTile(
+              title: Text("By plane"),
+              subtitle: Text("Viajar por plane"),
+              value: Transportation.plane,
+              groupValue: selectedTransportation,
+              onChanged: (value) => setState(() {
+                selectedTransportation = Transportation.plane;
+              }),
+            ),
+
+            RadioListTile(
+              title: Text("By submarine"),
+              subtitle: Text("Viajar por submarine"),
+              value: Transportation.submarine,
+              groupValue: selectedTransportation,
+              onChanged: (value) => setState(() {
+                selectedTransportation = Transportation.submarine;
+              }),
+            ),
+          ],
+        ),
+
+        CheckboxListTile(
+          title: Text("wantsBreakfast?"),
+          value: wantsBreakfast,
+          onChanged: (value) => setState(() {
+            wantsBreakfast = !wantsBreakfast;
+          }),
+        ),
+
+        CheckboxListTile(
+          title: Text("wantsLunch?"),
+          value: wantsLunch,
+          onChanged: (value) => setState(() {
+            wantsLunch = !wantsLunch;
+          }),
+        ),
+
+        CheckboxListTile(
+          title: Text("wantsDinner?"),
+          value: wantsDinner,
+          onChanged: (value) => setState(() {
+            wantsDinner = !wantsDinner;
+          }),
+        ),
+      ],
+    );
+  }
+}